feat(cangjie_tools): sync from upstream v1.1.0-alpha.68

+++ cangjie-language-server/src/languageserver/index/IndexDatabase.cpp
@@ -1441,15 +1441,15 @@ dberr_no OpenIndexDatabase(IndexDatabase &db, const std::string &file,
 {
     return db.Initialize([dbfile = file, dbOpts = std::move(opts)] {
         if (dbOpts.openInMemory) {
-            return (sqldb::memdb)(
-                dbfile, sqldb::OpenURI | (dbOpts.openReadOnly ? sqldb::OpenReadOnly
-                                                          : (sqldb::OpenCreate |
-                                                             sqldb::OpenReadWrite)));
+            return (sqldb::memdb)(dbfile,
+                (unsigned int)sqldb::OpenURI |
+                    (dbOpts.openReadOnly ? (unsigned int)sqldb::OpenReadOnly
+                                         : ((unsigned int)sqldb::OpenCreate | (unsigned int)sqldb::OpenReadWrite)));
         } else {
-            return (sqldb::open)(
-                dbfile, sqldb::OpenURI | (dbOpts.openReadOnly ? sqldb::OpenReadOnly
-                                                        : (sqldb::OpenCreate |
-                                                           sqldb::OpenReadWrite)));
+            return (sqldb::open)(dbfile,
+                (unsigned int)sqldb::OpenURI |
+                    (dbOpts.openReadOnly ? (unsigned int)sqldb::OpenReadOnly
+                                         : ((unsigned int)sqldb::OpenCreate | (unsigned int)sqldb::OpenReadWrite)));
         }
     });
 }
